Change to ModelViewSet for LegalBasis API View
This allows create/retrieve/update/delete operations automatically.

modified:   server/apps/main/migrations/0001_initial.py
the initial migration for the main app has been regenerated, so you might want
to migrate down to zero and back up to ensure your database tables are in sync.
The change adds a unique constraint to the Consent model's name field. This is
to allow using the SlugRelatedField serializer in the `LegalBasisSerializer`