Poller management command for DirectoryFormsAPI
This is a management command that can poll the forms api for new submissions
that indicate a user's acceptance or rejection of email_marketing  (stored in
the email_contact_consent) field.

This also provides a template for other types of message in the activity stream
such as Aventri signups.

To make lookup of email address work if someone entered TEST@test.com vs
test@test.com the LegalBasis model's email field has been changed to a case
insensitive email field.

Note the initial migration of the `main` app has been regenerated so do:
./manage.py migrate main zero
./manage.py migrate

before trying to run the server after this commit