fix: add ordering to paginated response
The data-flow pipeline is failing due to primary key duplication. It's
suspected that the lack of ordering on paginated responses causes this - data
can be duplicated between pages because no global ordering is enforced.

This would be worse when there are changes to legal basis during the run of a
pipeline, and we're now making more changes to legal basis because of recent
fixes to the Forms API poller.